# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgentFuel API - BSC Agent Payment Layer
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Service {
  id: number;
  ownerAddress: string;
  name: string;
  description: string;
  endpoint: string;
  price: string;
  currency: string;
  stakeRequired: string;
  successRate: number;
  avgLatency: number;
  totalCalls: number;
  active: boolean;
  createdAt: string;
}

export interface CreateServiceInput {
  ownerAddress: string;
  name: string;
  description: string;
  endpoint: string;
  price: string;
  currency: string;
  stakeRequired: string;
}

export interface Provider {
  id: number;
  address: string;
  stakedFuel: string;
  reputationScore: number;
  slashCount: number;
  totalEarnings: string;
  createdAt: string;
}

export interface CreateProviderInput {
  address: string;
}

export type QuoteInputInput = { [key: string]: unknown };

export interface QuoteInput {
  serviceId: string;
  input?: QuoteInputInput;
}

export interface QuoteResult {
  ok: boolean;
  serviceId: string;
  price: string;
  token: string;
  stakeRequired: string;
  paymentHeader: string;
}

export interface Receipt {
  id: number;
  serviceId: number;
  payer: string;
  provider: string;
  token: string;
  amount: string;
  txHash?: string | null;
  requestHash?: string | null;
  responseHash?: string | null;
  success: boolean;
  latencyMs: number;
  createdAt: string;
}

export interface ErrorResponse {
  error: string;
}

export type ListReceiptsParams = {
  payer?: string;
  provider?: string;
  serviceId?: number;
};