# lib/api-zod/src/generated/types/createServiceInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgentFuel API - BSC Agent Payment Layer
 * OpenAPI spec version: 0.1.0
 */

export interface CreateServiceInput {
  ownerAddress: string;
  name: string;
  description: string;
  endpoint: string;
  price: string;
  currency: string;
  stakeRequired: string;
}