# lib/api-zod/src/generated/types/provider.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgentFuel API - BSC Agent Payment Layer
 * OpenAPI spec version: 0.1.0
 */

export interface Provider {
  id: number;
  address: string;
  stakedFuel: string;
  reputationScore: number;
  slashCount: number;
  totalEarnings: string;
  createdAt: string;
}