# lib/api-zod/src/generated/types/quoteInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgentFuel API - BSC Agent Payment Layer
 * OpenAPI spec version: 0.1.0
 */
import type { QuoteInputInput } from "./quoteInputInput";

export interface QuoteInput {
  serviceId: string;
  input?: QuoteInputInput;
}