# lib/api-zod/src/generated/types/quoteResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgentFuel API - BSC Agent Payment Layer
 * OpenAPI spec version: 0.1.0
 */

export interface QuoteResult {
  ok: boolean;
  serviceId: string;
  price: string;
  token: string;
  stakeRequired: string;
  paymentHeader: string;
}