# lib/api-zod/src/generated/types/receipt.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgentFuel API - BSC Agent Payment Layer
 * OpenAPI spec version: 0.1.0
 */

export interface Receipt {
  id: number;
  serviceId: number;
  payer: string;
  provider: string;
  token: string;
  amount: string;
  txHash?: string | null;
  requestHash?: string | null;
  responseHash?: string | null;
  success: boolean;
  latencyMs: number;
  createdAt: string;
}